move firebase ui config to config.js
+ remove comments from index.html
