+++ src/index.js
@@ -1,14 +1,13 @@
 // using ES6 imports:
 // full qualified path for webpack esm style
-import { firebaseConfig } from '../auth/config.js'
-import { getAuth, onAuthStateChanged, GoogleAuthProvider, EmailAuthProvider } from "firebase/auth"
+import { firebaseConfig, firebaseUiConfig } from '../auth/config.js'
 import { getDatabase } from "firebase/database"
 
 import * as firebaseui from "firebaseui"
 import 'firebaseui/dist/firebaseui.css'
 
 import { handleSignedInUser, handleSignedOutUser } from "./view/userview.js"
-import { writeUserData, deleteAccount } from "./model/userdbmapper.js"
+import { deleteAccount } from "./model/userdbmapper.js"
 
 import { tasksViewFactory, listsViewFactory, itemsCreationView } from './view/view.js'
 import { tasksModelFactory, listsModelFactory, itemsCreationModel } from './model/model.js'
@@ -39,43 +38,6 @@ window.addEventListener('load', initApp);
 const firebaseApp = firebase.initializeApp(firebaseConfig);
 const firebaseUi = new firebaseui.auth.AuthUI(firebaseApp.auth());
 const firebaseDatabase = getDatabase(firebaseApp);
-
-const firebaseUiConfig =
-{
-  callbacks: {
-    signInSuccessWithAuthResult: function (authResult, redirectUrl) {
-      // User successfully signed in.
-      // Return type determines whether we continue the redirect automatically
-      // or whether we leave that to developer to handle.
-      if (authResult.user) 
-      {
-        handleSignedInUser(authResult.user);
-      }  
-
-      writeUserData(firebaseDatabase, authResult.user.uid, authResult.user.displayName, authResult.user.email, null)
-      document.getElementById('container').removeAttribute('hidden')
-      // Do not redirect.
-      return false;
-    },
-    uiShown: function () {
-      // The widget is rendered.
-      // Hide the loader.
-      document.getElementById('loader').style.display = 'none'
-    }
-  },
-  // Will use popup for IDP Providers sign-in flow instead of the default, redirect.
-  signInFlow: 'popup',
-  // signInSuccessUrl: 'https://todo-extended.web.app/',
-  signInOptions: [
-    // Leave the lines as is for the providers you want to offer your users.
-    GoogleAuthProvider.PROVIDER_ID,
-    EmailAuthProvider.PROVIDER_ID
-  ],
-  // Terms of service url.
-  tosUrl: '<your-tos-url>',
-  // Privacy policy url.
-  privacyPolicyUrl: '<your-privacy-policy-url>'
-}
 
 // The start method will wait until the DOM is loaded.
 firebaseUi.start('#firebaseui-auth-container', firebaseUiConfig)
